% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/newsfreq.R
\name{autoplot.newsfreq}
\alias{autoplot.newsfreq}
\title{Autoplot for \code{newsfreq} objects}
\usage{
\method{autoplot}{newsfreq}(data, breaks = "year", label_format = "\%Y",
  value = "count", add_point = FALSE)
}
\arguments{
\item{data}{a \code{newsfreq} object}

\item{breaks}{A character string, containing one of "\code{day}", "\code{week}",
"\code{month}", "\code{quarter}" or "\code{year}". This can optionally
be preceded by a (positive or negative) integer and a space, or followed
by "\code{s}". Passed to \code{scale_x_date}. Defaults to "\code{year}".}

\item{label_format}{passed to \code{scale_x_date}. Defaults to "\code{\%Y}"}

\item{value}{either "\code{count}" for raw article counts or "\code{percent}"
Defaults to "\code{count}".}

\item{add_point}{add points to the lines? Default "\code{FALSE}" (do not plot points)}
}
\value{
\code{ggplot} object
}
\description{
Quick method with sane defaults for generating a \code{ggplot}
       plot for \code{news_search} results.
}

