% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/newsfreq.R
\name{news_search}
\alias{is.newsfreq}
\alias{news_search}
\title{Search newsfreq.com API}
\usage{
news_search(keywords = NA, target = "", date_from = (Sys.Date() - 1),
  date_to = (Sys.Date() - 1), source = "", summarize = "monthly")

is.newsfreq(x)
}
\arguments{
\item{keywords}{search term(s) to query (see Details for specifics)}

\item{target}{news article component to search in (see Details for valid values and their meaning)}

\item{date_from}{start date for the search (<= \code{date_to} and not current day or future date).
Defaults to yesterday. See Details for information on date formatting.}

\item{date_to}{end date for search (>= \code{date_from} and not current day). Defaults to yesterday.
See Details for information on date formatting.}

\item{source}{filter search by news source. You can filter your search by news organization.
Entering 'Fox News' or 'The Boston Globe' will search only articles from sources matching that name.
Full list available on \href{http://bit.ly/newsbanksources}{NewsBank}}

\item{summarize}{Either "\code{monthly}" or "\code{annual}"}
}
\value{
\code{data.frame} (with additional class of \code{newsfreq}) of search
    results. If \code{summarize} is "\code{monthly}",
    then the results will also include \code{year}, \code{month} and
    \code{month_abb}. For "\code{annual}", only \code{year} will be added to
    the results.
}
\description{
\code{news_search} provides an interface to the \code{newsfreq.com} news keyword search API
}
\details{
\code{newsfreq.com}'s interface shows you the frequency that given keywords appear in
    American News Media and can be used to look at trends in news reporting.
    \code{news_search} provides programmatic access to the search function, returning results in an R data frame where
    additional analyses can be performed or data can be visualized.\cr
    \cr
    You can use boolean operators \code{AND} and \code{OR} with parentheses to specify multiple terms in \code{keywords}.\cr
    \cr
    The \code{target} parameter controls which "field" in news stories are queried. Valid values are:
    \itemize{
      \item \code{""} Search all article fields (the default)
      \item \code{lead} Search in article lead paragraph
      \item \code{title} Search in article headlines
      \item \code{topics} Search in terms identified as index terms by the article provider
      \item \code{author} Search for articles by a particular author
      \item \code{section} Search for articles appearing in a particular news section
      \item \code{source} Search for articles from a particular news source
    }
    Search dates must not be in the future and must also not be on the current day (they won't be in the API database).
    They can be an atomic character vector in a format \code{as.Date} will recognize or
    anything that can be coerced to a \code{Date} class.
}
\note{
The "\code{percent}" value is a per-annum calculation and will only be calculated
    and added to the results of searches that are summarized \code{monthly} and span full years.
    It is at best a crude way to normalize the results since the number of news sources
    (and, hence, articles) changes over time.
}
\examples{
\dontrun{
news_search("data breach", date_from="2014-01-01", date_to="2014-12-31")
##     date_from    date_to year month_abb month count search_date search_terms        pct
## 1  2014-01-01 2014-01-31 2014       Jan    01  3963  2015-02-02  data breach 0.12539155
## 2  2014-02-01 2014-02-28 2014       Feb    02  2856  2015-02-02  data breach 0.09036545
## 3  2014-03-01 2014-03-31 2014       Mar    03  2589  2015-02-02  data breach 0.08191742
## 4  2014-04-01 2014-04-30 2014       Apr    04  2170  2015-02-02  data breach 0.06866002
## 5  2014-05-01 2014-05-31 2014       May    05  2680  2015-02-02  data breach 0.08479671
## 6  2014-06-01 2014-06-30 2014       Jun    06  1973  2015-02-02  data breach 0.06242683
## 7  2014-07-01 2014-07-31 2014       Jul    07  1962  2015-02-02  data breach 0.06207879
## 8  2014-08-01 2014-08-31 2014       Aug    08  2585  2015-02-02  data breach 0.08179086
## 9  2014-09-01 2014-09-30 2014       Sep    09  3326  2015-02-02  data breach 0.10523651
## 10 2014-10-01 2014-10-31 2014       Oct    10  2862  2015-02-02  data breach 0.09055529
## 11 2014-11-01 2014-11-30 2014       Nov    11  2473  2015-02-02  data breach 0.07824711
## 12 2014-12-01 2014-12-31 2014       Dec    12  2166  2015-02-02  data breach 0.06853346
}
}

