% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/newsfreq-package.R
\docType{package}
\name{newsfreq}
\alias{newsfreq}
\alias{newsfreq-package}
\title{A package to search for (via \url{http://newsfreq.com/}) keyword frequencies in news articles}
\description{
A package to search for (via \url{http://newsfreq.com/}) keyword frequencies in news articles
}
\author{
Bob Rudis (@hrbrmstr)
}

