% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/newsfreq.R
\name{news_search}
\alias{news_search}
\title{Search newsfreq.com API}
\usage{
news_search(keywords = NA, target = "", date_from = (Sys.Date() - 1),
  date_to = (Sys.Date() - 1), source = "", summarize = "monthly")
}
\arguments{
\item{keywords}{search term(s) to query (see Details for specifics)}

\item{target}{news article component to search in (see Details for valid values and their meaning)}

\item{date_from}{start date for the search (<= \code{date_to} and not current day or future date).
Defaults to yesterday. See Details for information on date formatting.}

\item{date_to}{end date for search (>= \code{date_from} and not current day). Defaults to yesterday.
See Details for information on date formatting.}

\item{source}{filter search by news source. You can filter your search by news organization.
Entering 'Fox News' or 'The Boston Globe' will search only articles from sources matching that name.
Full list available on \href{http://bit.ly/newsbanksources}{NewsBank}}

\item{summarize}{Either "\code{monthly}" or "\code{annual}"}
}
\value{
\code{data.frame} of search results. If \code{summarize} is "\code{monthly}",
    then the \code{data.frame} will also include \code{year}, \code{month} and
    \code{month_abb}. For "\code{annual}", only \code{year} will be added to the results.
}
\description{
\code{news_search} provides an interface to the \code{newsfreq.com} news keyword search API
}
\details{
\code{newsfreq.com}'s interface shows you the frequency that given keywords appear in
    American News Media and can be used to look at trends in news reporting.
    \code{news_search} provides programmatic access to the search function, returning results in an R data frame where
    additional analyses can be performed or data can be visualized.\cr
    \cr
    You can use boolean operators \code{AND} and \code{OR} with parentheses to specify multiple terms in \code{keywords}.\cr
    \cr
    The \code{target} parameter controls which "field" in news stories are queried. Valid values are:
    \itemize{
      \item \code{""} Search all article fields (the default)
      \item \code{lead} Search in article lead paragraph
      \item \code{title} Search in article headlines
      \item \code{topics} Search in terms identified as index terms by the article provider
      \item \code{author} Search for articles by a particular author
      \item \code{section} Search for articles appearing in a particular news section
      \item \code{source} Search for articles from a particular news source
    }
    Search dates must not be in the future and must also not be on the current day (they won't be in the API database). \cr
    They can be an atomic character vector in \code{mm/dd/YYYY} format or anything that can be coerced to a \code{Date} class.
}
\examples{
\dontrun{
news_search("data breach", date_from="2014-01-01", date_to="2014-12-31")
##     date_from    date_to count search_date search_terms
## 1  2014-01-01 2014-01-31  3963  2015-01-31  data breach
## 2  2014-02-01 2014-02-28  2856  2015-01-31  data breach
## 3  2014-03-01 2014-03-31  2589  2015-01-31  data breach
## 4  2014-04-01 2014-04-30  2170  2015-01-31  data breach
## 5  2014-05-01 2014-05-31  2680  2015-01-31  data breach
## 6  2014-06-01 2014-06-30  1973  2015-01-31  data breach
## 7  2014-07-01 2014-07-31  1962  2015-01-31  data breach
## 8  2014-08-01 2014-08-31  2585  2015-01-31  data breach
## 9  2014-09-01 2014-09-30  3326  2015-01-31  data breach
## 10 2014-10-01 2014-10-31  2862  2015-01-31  data breach
## 11 2014-11-01 2014-11-30  2473  2015-01-31  data breach
## 12 2014-12-01 2014-12-31  2166  2015-01-31  data breach
}
}

